refactor(client): extract AnalyticsTab from ArtistPage into memoized component (C26)
Move the analytics TabsContent verbatim into components/artist/AnalyticsTab
(React.memo). Shared pure helpers getQualityColor/getReleaseTypeBadge/
getStatusBadge moved to components/artist/artistPageUtils and imported by
both the page and the tab. No behavior change.
